requirements.txt
Pillow>=11.0.0
google-genai~=1.55.0
python-dotenv~=1.2.1
opencv-python~=4.12.0.88
